Refactor: Apply finial power bonus directly in staff
