+++ src/main/java/com/github/ars_zero/common/item/AbstractSpellStaff.java
@@ -1,17 +1,23 @@
 package com.github.ars_zero.common.item;
 
+import com.github.ars_zero.ArsZero;
 import com.github.ars_zero.client.gui.ArsZeroStaffGUI;
 import com.github.ars_zero.client.gui.AbstractMultiPhaseCastDeviceScreen;
 import com.github.ars_zero.common.spell.MultiPhaseCastContext;
+import com.alexthw.sauce.registry.ModRegistry;
 import com.hollingsworth.arsnouveau.api.registry.SpellCasterRegistry;
+import com.hollingsworth.arsnouveau.api.spell.Spell;
 import com.hollingsworth.arsnouveau.api.spell.SpellCaster;
 import com.hollingsworth.arsnouveau.api.spell.SpellTier;
 import net.minecraft.client.Minecraft;
 import net.minecraft.network.chat.Component;
+import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResultHolder;
 import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.entity.ai.attributes.AttributeInstance;
+import net.minecraft.world.entity.ai.attributes.AttributeModifier;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.core.component.DataComponents;
 import net.minecraft.nbt.CompoundTag;
@@ -159,6 +165,43 @@ public abstract class AbstractSpellStaff extends AbstractMultiPhaseCastDevice im
     
     public static boolean hasFinial(ItemStack stack) {
         return getFinial(stack) != null;
+    }
+    
+    private static final ResourceLocation FINIAL_POWER_BONUS_ID = ArsZero.prefix("finial_power_bonus");
+    
+    @Override
+    protected boolean checkManaAndCast(Player player, ItemStack stack, Spell spell, Phase phase) {
+        AttributeModifier finialModifier = null;
+        AttributeInstance powerAttribute = null;
+        
+        String finial = getFinial(stack);
+        if (finial != null && !player.level().isClientSide) {
+            powerAttribute = getPowerAttributeForFinial(player, finial);
+            if (powerAttribute != null) {
+                finialModifier = new AttributeModifier(
+                    FINIAL_POWER_BONUS_ID,
+                    1.0,
+                    AttributeModifier.Operation.ADD_VALUE
+                );
+                powerAttribute.addTransientModifier(finialModifier);
+            }
+        }
+        
+        try {
+            return super.checkManaAndCast(player, stack, spell, phase);
+        } finally {
+            if (finialModifier != null && powerAttribute != null) {
+                powerAttribute.removeModifier(FINIAL_POWER_BONUS_ID);
+            }
+        }
+    }
+    
+    private static AttributeInstance getPowerAttributeForFinial(Player player, String finial) {
+        return switch (finial) {
+            case "fire" -> player.getAttribute(ModRegistry.FIRE_POWER);
+            case "water" -> player.getAttribute(ModRegistry.WATER_POWER);
+            default -> null;
+        };
     }
 }
